Phase out legacy experiment updated hook
