+++ packages/back-end/src/services/experiments.ts
@@ -29,8 +29,6 @@ import {
 import { SegmentInterface } from "../../types/segment";
 import { ExperimentInterface } from "../../types/experiment";
 import { PastExperiment } from "../../types/past-experiments";
-import { queueWebhook } from "../jobs/webhooks";
-import { queueCDNInvalidate } from "../jobs/cacheInvalidate";
 import { promiseAllChunks } from "../util/promise";
 import { findDimensionById } from "../models/DimensionModel";
 import { getValidDate } from "../util/dates";
@@ -46,7 +44,6 @@ import {
 } from "../../types/organization";
 import { StatsEngine } from "../../types/stats";
 import { logger } from "../util/logger";
-import { getSDKPayloadKeys } from "../util/features";
 import { DataSourceInterface } from "../../types/datasource";
 import {
   ApiExperiment,
@@ -583,27 +580,6 @@ export async function processPastExperiments(
   return Array.from(experimentMap.values()).filter(
     (e) => e.numVariations > 1 && e.numVariations < 10
   );
-}
-
-export async function experimentUpdated(
-  experiment: ExperimentInterface,
-  previousProject: string = ""
-) {
-  const payloadKeys = getSDKPayloadKeys(
-    new Set(["dev", "production"]),
-    new Set(["", previousProject || "", experiment.project || ""])
-  );
-
-  // fire the webhook:
-  await queueWebhook(experiment.organization, payloadKeys, false);
-
-  // invalidate the CDN
-  await queueCDNInvalidate(experiment.organization, (key) => {
-    // Which url to invalidate depends on the type of experiment
-    return experiment.implementation === "visual"
-      ? `/js/${key}.js`
-      : `/config/${key}`;
-  });
 }
 
 function getExperimentMetric(
